requirements.txt
openai
langchain
mysql-connector-python
python-dotenv
streamlit     # only if using UI
